ref(cypress): extract rules
