+++ packages/schematics/src/cypress/rules/utils.ts
@@ -1,0 +1,3 @@
+export function getE2eProjectNameForApp(appName: string) {
+  return `${appName}-e2e`;
+}
