+++ packages/schematics/src/lib/index.ts
@@ -1,5 +1,7 @@
+export * from './angular-json';
+export * from './angular-json-project';
 export * from './package-json';
 export * from './rules/install-dependencies';
-export * from './rules/remove-dependencies';
 export * from './rules/install-options';
+export * from './rules/remove-dependencies';
 export * from './rules/warn-against-competing-configuration';
